+++ Impl/arangodb/run_tasks/Task11_1.js
@@ -1,3 +1,4 @@
-db.T11_temp.drop()
+if (db._collection("T11_temp") != null) db.T11_temp.drop();
+
 db._create("T11_temp")
