+++ Impl/arangodb/run_tasks/Task8_1.js
@@ -1,8 +1,8 @@
 db._useDatabase("Healthcare")
 
-db.drug_temp.drop()
-db.target_temp.drop()
-db.has_bond.drop()
+if (db._collection("drug_temp") != null) db.drug_temp.drop();
+if (db._collection("target_temp") != null) db.target_temp.drop();
+if (db._collection("has_bond") != null) db.has_bond.drop();
 
 db._create("drug_temp")
 db._create("target_temp")
